+++ cli/actions/importKeys.go
@@ -27,6 +27,7 @@ import (
 
 	"github.com/NethermindEth/sedge/configs"
 	"github.com/NethermindEth/sedge/internal/images/validator-import/lighthouse"
+	"github.com/NethermindEth/sedge/internal/images/validator-import/prysm"
 	"github.com/NethermindEth/sedge/internal/images/validator-import/teku"
 	"github.com/NethermindEth/sedge/internal/pkg/commands"
 	"github.com/NethermindEth/sedge/internal/pkg/services"
@@ -51,6 +52,7 @@ type ImportValidatorKeysOptions struct {
 	GenerationPath  string
 	ContainerTag    string
 	CustomConfig    ImportValidatorKeysCustomOptions
+	Distributed     bool
 }
 type ImportValidatorKeysCustomOptions struct {
 	NetworkConfigPath string
@@ -100,19 +102,95 @@ func (s *sedgeActions) ImportValidatorKeys(options ImportValidatorKeysOptions) e
 	options.GenerationPath = absGenerationPath
 
 	if !isDefaultKeysPath(options.GenerationPath, options.From) {
-		defaultKeystorePath := filepath.Join(options.GenerationPath, "keystore")
-		log.Warnf("The keys path is not the default one, copying the keys to the default path %s", defaultKeystorePath)
-		copy.Copy(options.From, defaultKeystorePath)
+		if !options.Distributed {
+			defaultKeystorePath := filepath.Join(options.GenerationPath, "keystore")
+			log.Warnf("The keys path is not the default one, copying the keys to the default path %s", defaultKeystorePath)
+			copy.Copy(options.From, defaultKeystorePath)
+		}
+	}
+
+	if options.Distributed {
+		cwd, _ := os.Getwd()
+		charonPath := filepath.Join(cwd, ".charon")
+
+		if !isDefaultKeysPath(options.GenerationPath, options.From) {
+			charonPath = options.From
+			log.Infof("Copying the keys from %s", charonPath)
+			options.From = filepath.Join(options.GenerationPath, "keystore")
+		}
+		defaultCharonPath := filepath.Join(configs.DefaultAbsSedgeDataPath, ".charon")
+		// Copy the folder from charonPath to defaultCharonPath
+		log.Infof("Copying Charon contents to the default path %s", defaultCharonPath)
+		if err := os.MkdirAll(defaultCharonPath, 0o755); err != nil {
+			return err
+		}
+		if err := copy.Copy(charonPath, defaultCharonPath); err != nil {
+			return err
+		}
+		charonValidatorKeysPath := filepath.Join(charonPath, "validator_keys")
+		defaultKeystorePath := filepath.Join(configs.DefaultAbsSedgeDataPath, "keystore")
+		log.Infof("Copying the keys to the default path %s", defaultKeystorePath)
+		if err := os.MkdirAll(defaultKeystorePath, 0o755); err != nil {
+			return err
+		}
+
+		validatorKeysPath := filepath.Join(defaultKeystorePath, "validator_keys")
+		if err := os.MkdirAll(validatorKeysPath, 0o755); err != nil {
+			return err
+		}
+
+		depositDataPath := filepath.Join(charonPath, "deposit-data.json")
+		depositDataPathDest := filepath.Join(defaultKeystorePath, "deposit-data.json")
+		if err := copy.Copy(depositDataPath, depositDataPathDest); err != nil {
+			return err
+		}
+
+		files, err := os.ReadDir(charonValidatorKeysPath)
+		if err != nil {
+			log.Fatal(err)
+		}
+		len := len(files)
+		for i := 0; i < len/2; i++ {
+			keystorePath := filepath.Join(charonValidatorKeysPath, fmt.Sprintf("keystore-%d.json", i))
+			validatorPath := filepath.Join(validatorKeysPath, fmt.Sprintf("keystore-%d.json", i))
+			if err := copy.Copy(keystorePath, validatorPath); err != nil {
+				return err
+			}
+
+			keystoreTxtPath := filepath.Join(charonValidatorKeysPath, fmt.Sprintf("keystore-%d.txt", i))
+			keystorePasswordPath := filepath.Join(defaultKeystorePath, fmt.Sprintf("keystore-%d.txt", i))
+			if err := copy.Copy(keystoreTxtPath, keystorePasswordPath); err != nil {
+				return err
+			}
+		}
+		if options.ValidatorClient == "prysm" {
+			keystorePasswordPath := filepath.Join(defaultKeystorePath, "keystore_password.txt")
+			f, err := os.Create(keystorePasswordPath)
+			if err != nil {
+				return err
+			}
+			f.WriteString("prysm-validator-secret")
+			defer f.Close()
+		}
 	}
 
 	var ctID string
 	switch options.ValidatorClient {
 	case "prysm":
-		prysmCtID, err := setupPrysmValidatorImportContainer(s.dockerClient, s.dockerServiceManager, options)
-		if err != nil {
-			return err
+		prysmCtID := ""
+		if options.Distributed {
+			prysmCtID, err = setupPrysmValidatorImportContainerDV(s.dockerClient, s.commandRunner, s.dockerServiceManager, options)
+			if err != nil {
+				return err
+			}
+			ctID = prysmCtID
+		} else {
+			prysmCtID, err := setupPrysmValidatorImportContainer(s.dockerClient, s.dockerServiceManager, options)
+			if err != nil {
+				return err
+			}
+			ctID = prysmCtID
 		}
-		ctID = prysmCtID
 	case "nimbus":
 		nimbusCtID, err := setupNimbusValidatorImport(s.dockerClient, s.dockerServiceManager, options)
 		if err != nil {
@@ -143,7 +221,11 @@ func (s *sedgeActions) ImportValidatorKeys(options ImportValidatorKeysOptions) e
 	log.Info("Importing validator keys")
 	var runErr error
 	if options.ValidatorClient == "nimbus" {
-		runErr = runAndWaitImportKeysNimbus(s.dockerClient, s.dockerServiceManager, ctID)
+		if !options.Distributed {
+			runErr = runAndWaitImportKeysNimbus(s.dockerClient, s.dockerServiceManager, ctID)
+		} else {
+			runErr = runAndWaitImportKeys(s.dockerClient, s.dockerServiceManager, ctID)
+		}
 	} else {
 		runErr = runAndWaitImportKeys(s.dockerClient, s.dockerServiceManager, ctID)
 	}
@@ -260,17 +342,47 @@ func setupNimbusValidatorImport(dockerClient client.APIClient, dockerServiceMana
 	} else {
 		cmd = append(cmd, "--network="+options.Network)
 	}
+	containerConfig := &container.Config{
+		Image:        validatorImage,
+		Cmd:          cmd,
+		AttachStdin:  true,
+		AttachStderr: true,
+		AttachStdout: true,
+		OpenStdin:    true,
+		Tty:          true,
+	}
+	if options.Distributed {
+		containerConfig = &container.Config{
+			Image: validatorImage,
+			Entrypoint: []string{
+				"sh", "-c", `
+				#!/usr/bin/env bash
+				set -e
+				tmpkeys="/keystore/validator_keys/tmpkeys"
+				mkdir -p ${tmpkeys}
+				for f in /keystore/validator_keys/keystore-*.json; do
+					echo "Importing key ${f}"
+					pwdfile="/keystore/$(basename "$f" .json).txt"
+					password=$(cat ${pwdfile})
+					echo "Using password file ${pwdfile}"
+					echo "Using password ${password}"
+					cp "${f}" "${tmpkeys}"
+					# Import keystore with password.
+						echo "$password" | \
+						/home/user/nimbus_beacon_node deposits import \
+						--data-dir=/data \
+						${tmpkeys}
+					filename="$(basename ${f})"
+  					rm "${tmpkeys}/${filename}"
+				done
+			`,
+			},
+		}
+	}
+
 	log.Debugf("Creating %s container", validatorImportCtName)
 	ct, err := dockerClient.ContainerCreate(context.Background(),
-		&container.Config{
-			Image:        validatorImage,
-			Cmd:          cmd,
-			AttachStdin:  true,
-			AttachStderr: true,
-			AttachStdout: true,
-			OpenStdin:    true,
-			Tty:          true,
-		},
+		containerConfig,
 		&container.HostConfig{
 			Mounts:      mounts,
 			VolumesFrom: []string{consensusCtName, validatorCtName},
@@ -330,11 +442,33 @@ func setupLodestarValidatorImport(dockerClient client.APIClient, dockerServiceMa
 		cmd = append(cmd, "--preset", preset)
 	}
 	log.Debugf("Creating %s container", validatorImportCtName)
-	ct, err := dockerClient.ContainerCreate(context.Background(),
-		&container.Config{
+	containerConfig := &container.Config{
+		Image: validatorImage,
+		Cmd:   cmd,
+	}
+	if options.Distributed {
+		containerConfig = &container.Config{
 			Image: validatorImage,
-			Cmd:   cmd,
-		},
+			Entrypoint: []string{
+				"sh", "-c", `
+				#!/bin/sh
+				set -e
+				for f in /keystore/validator_keys/keystore-*.json; do
+					echo "Importing key ${f}"
+					pwdfile="/keystore/$(basename "$f" .json).txt"
+					echo "Using password file ${pwdfile}"
+					# Import keystore with password.
+					node /usr/app/packages/cli/bin/lodestar validator import \
+						--dataDir="/data" \
+						--importKeystores="$f" \
+						--importKeystoresPassword="${pwdfile}"
+				done
+			`,
+			},
+		}
+	}
+	ct, err := dockerClient.ContainerCreate(context.Background(),
+		containerConfig,
 		&container.HostConfig{
 			Mounts:      mounts,
 			VolumesFrom: []string{validatorCtName},
@@ -513,6 +647,57 @@ func runAndWaitImportKeys(dockerClient client.APIClient, dockerServiceManager Do
 			return exitErr
 		}
 	}
+}
+
+func setupPrysmValidatorImportContainerDV(dockerClient client.APIClient, commandRunner commands.CommandRunner, serviceManager DockerServiceManager, options ImportValidatorKeysOptions) (string, error) {
+	var (
+		validatorCtName       = services.ContainerNameWithTag(services.DefaultSedgeValidatorClient, options.ContainerTag)
+		validatorImportCtName = services.ContainerNameWithTag(services.ServiceCtValidatorImport, options.ContainerTag)
+	)
+	// Init build context
+	contextDir, err := prysm.InitContext()
+	if err != nil {
+		return "", err
+	}
+	// Build image
+	buildCmd := commandRunner.BuildDockerBuildCMD(commands.DockerBuildOptions{
+		Path: contextDir,
+		Tag:  "sedge/prysm-import-teku",
+		Args: map[string]string{
+			"NETWORK":       options.Network,
+			"PRYSM_VERSION": configs.ClientImages.Validator.Prysm.String(),
+		},
+	})
+	log.Infof(configs.RunningCommand, buildCmd.Cmd)
+	if _, _, err := commandRunner.RunCMD(buildCmd); err != nil {
+		return "", err
+	}
+	// Mounts
+	mounts := []mount.Mount{
+		{
+			Type:   mount.TypeBind,
+			Source: options.From,
+			Target: "/keystore",
+		},
+	}
+	log.Debugf("Creating %s container", validatorImportCtName)
+	containerConfig := &container.Config{
+		Image: "sedge/prysm-import-teku",
+	}
+	ct, err := dockerClient.ContainerCreate(context.Background(),
+		containerConfig,
+		&container.HostConfig{
+			Mounts:      mounts,
+			VolumesFrom: []string{validatorCtName},
+		},
+		&network.NetworkingConfig{},
+		&v1.Platform{},
+		validatorImportCtName,
+	)
+	if err != nil {
+		return "", err
+	}
+	return ct.ID, nil
 }
 
 // runAndWaitImportKeysNimbus starts the container in interactive mode and waits for it to finish.