+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+##Added
+- Support Mainnet for Lido CSM setup.
+
 ## [v1.6.0] - 2024-10-18
 
 ### Added